# docs/reference/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "cookie": {
        "in": "cookie",
        "name": "nexora_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "Nexora Platform Core API",
    "version": "1"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/memberships": {
      "get": {
        "operationId": "MembershipsController_list",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A cursor-bounded membership page."
          },
          "403": {
            "description": "Membership read permission denied."
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List active-workspace memberships",
        "tags": [
          "Memberships"
        ]
      }
    },
    "/v1/memberships/me": {
      "delete": {
        "operationId": "MembershipsController_leave",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "maxProperties": 0,
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "204": {
            "description": "Active workspace membership left."
          },
          "403": {
            "description": "Workspace leave permission denied."
          },
          "409": {
            "description": "Workspace ownership or final membership is protected."
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Leave the active workspace",
        "tags": [
          "Memberships"
        ]
      }
    },
    "/v1/memberships/owner": {
      "put": {
        "operationId": "MembershipsController_transferWorkspaceOwner",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "currentPassword": {
                    "description": "Maximum 128 normalized Unicode code points and 512 UTF-8 bytes.",
                    "maxLength": 512,
                    "minLength": 1,
                    "type": "string",
                    "writeOnly": true
                  },
                  "membershipId": {
                    "format": "uuid",
                    "type": "string"
                  }
                },
                "required": [
                  "membershipId",
                  "currentPassword"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Workspace ownership transferred."
          },
          "400": {
            "description": "Transfer confirmation was invalid."
          },
          "403": {
            "description": "Ownership transfer permission denied."
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Transfer active-workspace operational ownership",
        "tags": [
          "Memberships"
        ]
      }
    },
    "/v1/memberships/{membershipId}": {
      "delete": {
        "operationId": "MembershipsController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "membershipId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Membership removed or not visible in the active workspace."
          },
          "403": {
            "description": "Membership removal permission denied."
          },
          "409": {
            "description": "Workspace ownership is protected."
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Remove a non-owner active-workspace membership",
        "tags": [
          "Memberships"
        ]
      }
    },
    "/v1/memberships/{membershipId}/role": {
      "patch": {
        "operationId": "MembershipsController_updateRole",
        "parameters": [
          {
            "in": "path",
            "name": "membershipId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "role": {
                    "enum": [
                      "ADMIN",
                      "MEMBER"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "role"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Membership role updated or unchanged."
          },
          "403": {
            "description": "Role update permission denied."
          },
          "409": {
            "description": "Workspace ownership is protected."
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Change a non-owner membership role",
        "tags": [
          "Memberships"
        ]
      }
    },
...
  },
  "servers": [],
...
}